requirements.txt
flask==2.0.1
Flask_Cors
langchain==0.0.267
openai==0.27.8
python-dotenv==0.19.0
azure-core==1.29.3
azure-ai-inference==0.1.0
azure-identity==1.12.0
cassio==0.1.3
chromadb==0.3.29
datasets==2.14.5
pypdf==3.9.0
python-dotenv==0.19.0
tiktoken==0.4.0
typing-extensions==4.5.0
numpy>=1.20.0
pandas>=1.3.0
tenacity==8.2.2
aiohttp==3.8.4
requests==2.28.2
